Added shorter weekday names for mobile and tablet
Mobile and Tablet views would overlap or get too close at certain
resolutions. For some reason when i added the second set of elemnts
in the header, the last saturday would wrap to the next row, so I
explicitly set a width to make it not wrap. this seems a little dirty
I'm open to suggestions on fixing it, but this works for now.
closes #3